CBOR serialization

+++ src/types/proof_of_time.py
@@ -1,13 +1,13 @@
 from src.util.streamable import streamable, StreamableList
 from src.types.sized_bytes import bytes32
 from src.types.classgroup import ClassgroupElement
-from src.util.ints import uint8
+from src.util.ints import uint8, uint64
 
 
 @streamable
 class ProofOfTimeOutput:
     challenge_hash: bytes32
-    number_of_iterations: int
+    number_of_iterations: uint64
     output: ClassgroupElement
